move provider names under providers pkg
