+++ pkg/autodiscovery/configresolver/configresolver.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/integration"
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/listeners"
-	"github.com/DataDog/datadog-agent/pkg/autodiscovery/providernames"
+	"github.com/DataDog/datadog-agent/pkg/autodiscovery/providers/names"
 )
 
 type variableGetter func(key []byte, svc listeners.Service) ([]byte, error)
@@ -95,7 +95,7 @@ func Resolve(tpl integration.Config, svc listeners.Service) (integration.Config,
 
 	// Ignore the config from file if it's overridden by an empty config
 	// or by a different config for the same check
-	if tpl.Provider == providernames.File && svc.GetCheckNames() != "" {
+	if tpl.Provider == names.File && svc.GetCheckNames() != "" {
 		checkNames := []string{}
 		err := json.Unmarshal([]byte(svc.GetCheckNames()), &checkNames)
 		if err != nil {
