+++ pkg/autodiscovery/providers/clusterchecks.go
@@ -8,9 +8,8 @@ package providers
 import (
 	"time"
 
-	"github.com/DataDog/datadog-agent/pkg/autodiscovery/providernames"
-
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/integration"
+	"github.com/DataDog/datadog-agent/pkg/autodiscovery/providers/names"
 	"github.com/DataDog/datadog-agent/pkg/clusteragent/clusterchecks/types"
 	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/util"
@@ -60,7 +59,7 @@ func (c *ClusterChecksConfigProvider) initClient() error {
 
 // String returns a string representation of the ClusterChecksConfigProvider
 func (c *ClusterChecksConfigProvider) String() string {
-	return providernames.ClusterChecks
+	return names.ClusterChecks
 }
 
 func (c *ClusterChecksConfigProvider) withinGracePeriod() bool {
