+++ pkg/autodiscovery/providers/endpointschecks.go
@@ -9,7 +9,7 @@ package providers
 
 import (
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/integration"
-	"github.com/DataDog/datadog-agent/pkg/autodiscovery/providernames"
+	"github.com/DataDog/datadog-agent/pkg/autodiscovery/providers/names"
 	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/util/clusteragent"
 	"github.com/DataDog/datadog-agent/pkg/util/kubernetes/kubelet"
@@ -44,7 +44,7 @@ func NewEndpointsChecksConfigProvider(cfg config.ConfigurationProviders) (Config
 
 // String returns a string representation of the EndpointsChecksConfigProvider
 func (c *EndpointsChecksConfigProvider) String() string {
-	return providernames.EndpointsChecks
+	return names.EndpointsChecks
 }
 
 // IsUpToDate updates the list of AD templates versions in the Agent's cache and checks the list is up to date compared to Kubernetes's data.
