+++ pkg/autodiscovery/providers/etcd.go
@@ -18,7 +18,7 @@ import (
 	"golang.org/x/net/context"
 
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/integration"
-	"github.com/DataDog/datadog-agent/pkg/autodiscovery/providernames"
+	"github.com/DataDog/datadog-agent/pkg/autodiscovery/providers/names"
 	"github.com/DataDog/datadog-agent/pkg/config"
 )
 
@@ -193,7 +193,7 @@ func (p *EtcdConfigProvider) IsUpToDate() (bool, error) {
 
 // String returns a string representation of the EtcdConfigProvider
 func (p *EtcdConfigProvider) String() string {
-	return providernames.Etcd
+	return names.Etcd
 }
 
 // hasTemplateFields verifies that a node array contains
