+++ pkg/autodiscovery/providers/names/provider_names.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2019 Datadog, Inc.
 
-package providernames
+package names
 
 // User-facing names for the config providers
 const (
