+++ pkg/autodiscovery/providers/zookeeper.go
@@ -18,7 +18,7 @@ import (
 	"github.com/samuel/go-zookeeper/zk"
 
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/integration"
-	"github.com/DataDog/datadog-agent/pkg/autodiscovery/providernames"
+	"github.com/DataDog/datadog-agent/pkg/autodiscovery/providers/names"
 	"github.com/DataDog/datadog-agent/pkg/config"
 )
 
@@ -55,7 +55,7 @@ func NewZookeeperConfigProvider(cfg config.ConfigurationProviders) (ConfigProvid
 
 // String returns a string representation of the ZookeeperConfigProvider
 func (z *ZookeeperConfigProvider) String() string {
-	return providernames.Zookeeper
+	return names.Zookeeper
 }
 
 // Collect retrieves templates from Zookeeper, builds Config objects and returns them
